Update README.md
This project focuses on conducting a security audit for Botium Toys. The audit aims to evaluate the company's IT assets, security policies, and compliance with industry standards. It identifies vulnerabilities, assesses risks, and proposes solutions to improve security measures. The project applies frameworks such as the NIST Cybersecurity Framework to ensure a systematic and effective evaluation of controls and practices. Deliverables include risk assessments, compliance reports, and recommendations for improving security posture.